+++ android/src/main/java/com/mapbox/mapboxgl/MapboxMapController.java
@@ -8,7 +8,6 @@ import android.Manifest;
 import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
-import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageManager;
 import android.content.res.AssetFileDescriptor;
 import android.content.res.AssetManager;
@@ -139,7 +138,7 @@ final class MapboxMapController
     MapboxMapOptions options,
     String accessToken,
     String styleStringInitial) {
-    Mapbox.getInstance(context, accessToken!=null ? accessToken : getAccessToken(context));
+    MapBoxUtils.getMapbox(context, accessToken);
     this.id = id;
     this.context = context;
     this.activityState = activityState;
@@ -154,23 +153,6 @@ final class MapboxMapController
       new MethodChannel(registrar.messenger(), "plugins.flutter.io/mapbox_maps_" + id);
     methodChannel.setMethodCallHandler(this);
     this.registrarActivityHashCode = registrar.activity().hashCode();
-  }
-
-  private static String getAccessToken(@NonNull Context context) {
-    try {
-      ApplicationInfo ai = context.getPackageManager().getApplicationInfo(context.getPackageName(), PackageManager.GET_META_DATA);
-      Bundle bundle = ai.metaData;
-      String token = bundle.getString("com.mapbox.token");
-      if (token == null || token.isEmpty()) {
-        throw new NullPointerException();
-      }
-      return token;
-    } catch (Exception e) {
-      Log.e(TAG, "Failed to find an Access Token in the Application meta-data. Maps may not load correctly. " +
-        "Please refer to the installation guide at https://github.com/tobrun/flutter-mapbox-gl#mapbox-access-token " +
-        "for troubleshooting advice." + e.getMessage());
-    }
-    return null;
   }
 
   @Override
